Add MERA clones of lm_eval/models scripts and modules
