+++ lm_eval/evaluator.py
@@ -403,6 +403,8 @@ def simple_evaluate(
         # add more detailed model info if available
         if isinstance(lm, lm_eval.models.huggingface.HFLM):
             results["config"].update(lm.get_model_info())
+        if isinstance(lm, lm_eval.models.huggingface_mera.HFLMMERA):
+            results["config"].update(lm.get_model_info())
         # add info about execution
         results["config"].update(
             {
