+++ lm_eval/models/hf_videolm_mera.py
@@ -5,14 +5,18 @@ from typing import List, Optional, Tuple, Union, Dict
 
 import numpy as np
 import torch
-import torchvision
-import torchcodec
+try:
+    import torchvision
+    import torchcodec
+except ModuleNotFoundError:
+    pass
+    
 from tqdm import tqdm
 from transformers import AutoModelForPreTraining, PreTrainedModel, AutoModelForVision2Seq
 
 from lm_eval.api.instance import Instance
 from lm_eval.api.registry import register_model
-from lm_eval.models.hf_vlms import HFMultimodalLM
+from lm_eval.models.hf_vlms_mera import HFMultimodalLMMERA
 from lm_eval.models.utils import (
     Collator,
     handle_stop_sequences,
@@ -24,8 +28,8 @@ from lm_eval.api.task import DEFAULT_VIDEO_PLACEHOLDER
 eval_logger = logging.getLogger(__name__)
 
 
-@register_model("hf_videolm")
-class HFVideoLlava(HFMultimodalLM):
+@register_model("hf-videolm-mera")
+class HFVideoLlavaMERA(HFMultimodalLMMERA):
     MULTIMODAL = True
     AUTO_MODEL_CLASS = AutoModelForVision2Seq
