+++ lm_eval/models/vllm_causallms_mera.py
@@ -36,6 +36,7 @@ try:
     from vllm.lora.request import LoRARequest
     from vllm.transformers_utils.tokenizer import get_tokenizer
     from vllm.utils import get_open_port
+    from vllm.transformers_utils.tokenizers.mistral import MistralTokenizer
 
     if parse_version(version("vllm")) >= parse_version("0.8.3"):
         from vllm.entrypoints.chat_utils import resolve_hf_chat_template
@@ -105,8 +106,8 @@ def _vllm_mp_worker(
     return None
 
 
-@register_model("vllm")
-class VLLM(TemplateLM):
+@register_model("vllm-mera")
+class VLLMMERA(TemplateLM):
     _DEFAULT_MAX_LENGTH = 2048
 
     def __init__(
@@ -344,12 +345,21 @@ class VLLM(TemplateLM):
     ) -> Union[List[int], List[List[int]]]:
         if not add_special_tokens:
             add_special_tokens = False or self.add_bos_token
-        encoding: Union[List[List[int]], List[int]] = self.tokenizer(
-            string,
-            add_special_tokens=add_special_tokens,
-            truncation=truncation,
-            return_attention_mask=False,
-        ).input_ids
+        if isinstance(string, str) or isinstance(string[0], str):
+            encoding: Union[List[List[int]], List[int]] = self.tokenizer(
+                string,
+                add_special_tokens=add_special_tokens,
+                truncation=truncation,
+                return_attention_mask=False,
+            ).input_ids
+        else:
+            # It must be a formatted chat history dict then
+            encoding: Union[List[List[int]], List[int]] = self.tokenizer.apply_chat_template(
+                string,
+                add_special_tokens=add_special_tokens,
+                truncation=truncation,
+                return_attention_mask=False,
+            )
 
         # left-truncate the encoded context to be at most `left_truncate_len` tokens long
         if left_truncate_len:
