+++ lm_eval/models/vllm_vlms_mera.py
@@ -14,9 +14,9 @@ from lm_eval.models.utils import (
     replace_placeholders,
     resize_image,
     undistribute,
+    content_image_to_content_image_url,
 )
-from lm_eval.models.vllm_causallms import VLLM
-
+from lm_eval.models.vllm_causallms_mera import VLLMMERA
 
 eval_logger = logging.getLogger(__name__)
 
@@ -26,6 +26,7 @@ try:
     from vllm import LLM, SamplingParams
     from vllm.lora.request import LoRARequest  # noqa: F401
     from vllm.transformers_utils.tokenizer import get_tokenizer  # noqa: F401
+    from vllm.transformers_utils.tokenizers.mistral import MistralTokenizer
 except ModuleNotFoundError:
     pass
 
@@ -33,8 +34,8 @@ except ModuleNotFoundError:
 DEFAULT_IMAGE_PLACEHOLDER = "<image>"
 
 
-@register_model("vllm-vlm")
-class VLLM_VLM(VLLM):
+@register_model("vllm-vlm-mera")
+class VLLM_VLM(VLLMMERA):
     MULTIMODAL = True
 
     def __init__(
@@ -68,13 +69,15 @@ class VLLM_VLM(VLLM):
             revision=revision,
             **kwargs,
         )
+        if "phi-3.5" in pretrained and self.batch_size > 1:
+            raise ValueError("For phi-3.5 models only batch_size = 1 is now supported")
         self.interleave = interleave
         self.max_images = max_images
         self.processor = transformers.AutoProcessor.from_pretrained(
             pretrained,
             revision=revision,
             trust_remote_code=trust_remote_code,
-        )
+        ) if not isinstance(self.tokenizer, MistralTokenizer) else None
         self.chat_applied: bool = False
 
     def tok_batch_multimodal_encode(
@@ -112,6 +115,7 @@ class VLLM_VLM(VLLM):
         generate: bool = False,
         max_tokens: int = None,
         stop: Optional[List[str]] = None,
+        pass_multimodal_args_to_chat_history: Optional[bool] = False,
         **kwargs,
     ):
         if generate:
@@ -130,7 +134,8 @@ class VLLM_VLM(VLLM):
                 model_args: dict, sampling_params, requests: List[List[dict]]
             ):
                 llm = LLM(**model_args)
-                return llm.generate(requests, sampling_params=sampling_params)
+                fn = llm.chat if pass_multimodal_args_to_chat_history else llm.generate
+                return fn(requests, sampling_params=sampling_params)
 
             # dispatch requests to all self.data_parallel_size workers, in interleaved fashion
             # interleaved important to balance context lengths across workers
@@ -143,15 +148,17 @@ class VLLM_VLM(VLLM):
             # flatten results
             return undistribute(results)
 
+        fn = self.model.chat if pass_multimodal_args_to_chat_history else self.model.generate
+
         if self.lora_request is not None:
-            outputs = self.model.generate(
+            outputs = fn(
                 requests,
                 sampling_params=sampling_params,
                 use_tqdm=True if self.batch_size == "auto" else False,
                 lora_request=self.lora_request,
             )
         else:
-            outputs = self.model.generate(
+            outputs = fn(
                 requests,
                 sampling_params=sampling_params,
                 use_tqdm=True if self.batch_size == "auto" else False,
@@ -218,9 +225,10 @@ class VLLM_VLM(VLLM):
     def generate_until(
         self, requests: List[Instance], disable_tqdm: bool = False
     ) -> List[str]:
-        if requests and len(requests[0].args) < 3:
-            # Fall back to non-multimodal generation.
-            return super().generate_until(requests=requests, disable_tqdm=disable_tqdm)
+        ### HERE ###
+        # if requests and len(requests[0].args) < 3:
+        #     # Fall back to non-multimodal generation.
+        #     return super().generate_until(requests=requests, disable_tqdm=disable_tqdm)
 
         res = []
 
@@ -231,8 +239,11 @@ class VLLM_VLM(VLLM):
             #   padded context length. this is useful to simplify the batching logic and more importantly to make
             #   automatic adaptive batches much much easier to implement
             # - any OOMs will happen right away rather than near the end
-            toks = self.tok_encode(x[0])
-            return -len(toks), x[0]
+            if "phi-3.5" in self.model_args["model"].lower():
+                toks = []
+            else:
+                toks = self.tok_encode(copy.deepcopy(x[0]))
+            return -len(toks), str(x[0])
 
         pbar = tqdm(
             total=len(requests),
@@ -253,7 +264,13 @@ class VLLM_VLM(VLLM):
         chunks = re_ords.get_batched(n=self.batch_size, batch_fn=None)
         eos = self.tokenizer.decode(self.eot_token_id)
         for chunk in chunks:
-            contexts, all_gen_kwargs, aux_arguments = zip(*chunk)
+            if len(chunk[0]) == 2:
+                contexts, all_gen_kwargs = zip(*chunk)
+                aux_arguments = []
+                pass_multimodal_args_to_chat_history = True
+            else:
+                pass_multimodal_args_to_chat_history = False
+                contexts, all_gen_kwargs, aux_arguments = zip(*chunk)
 
             visuals = [
                 [
@@ -264,7 +281,6 @@ class VLLM_VLM(VLLM):
                 ]
                 for arg in aux_arguments
             ]
-
             if not isinstance(contexts, list):
                 contexts = list(
                     contexts
@@ -290,14 +306,26 @@ class VLLM_VLM(VLLM):
 
             max_ctx_len = self.max_length - max_gen_toks
 
-            inputs = self.tok_batch_multimodal_encode(
-                contexts,
-                visuals,
-                left_truncate_len=max_ctx_len,
-            )
+            if not pass_multimodal_args_to_chat_history:
+                inputs = self.tok_batch_multimodal_encode(
+                    contexts,
+                    visuals,
+                    left_truncate_len=max_ctx_len,
+                )
+            else:
+                inputs = []
+                for chat_history in contexts:
+                    new_chat_history = []
+                    for message in chat_history:
+                        new_content = []
+                        for content in message["content"]:
+                            new_content.append(content_image_to_content_image_url(content, resize=(self.image_width, self.image_height, self.image_max_side)))
+                        message["content"] = new_content
+                        new_chat_history.append(message)
+                    inputs.append(new_chat_history)
 
             cont = self._multimodal_model_generate(
-                inputs, stop=until, generate=True, max_tokens=max_gen_toks, **kwargs
+                inputs, stop=until, generate=True, max_tokens=max_gen_toks, pass_multimodal_args_to_chat_history=pass_multimodal_args_to_chat_history, **kwargs
             )
 
             for output, context in zip(cont, contexts):
